+++ firmware/IIe-keyboard_emulator/src/utils.c
@@ -1,0 +1,6 @@
+#include "utils.h"
+#include "hardware/timer.h"
+
+uint32_t millis() {
+	return ((time_us_32() / 1000));
+}
