+++ firmware/IIe-keyboard_emulator/src/utils.h
@@ -1,0 +1,8 @@
+#ifndef __UTILS_H__
+#define __UTILS_H__
+
+#include <stdint.h>
+
+uint32_t millis();
+
+#endif
